ci: Autobuild packages for Ubuntu LTS as well
Since Debhelper 12 (used in Ubuntu Focal) doesn't have a nice way to
deal with spaces in filenames, and I didn#t want to manually copy files
in d/rules or use dh-exec to keep the packaging as simple as possible,
removing spaces from the path was the easiest option.
Aside from that, building the Ubuntu package is straightforward: Just
use "focal" as suite instead of "bullseye".